requirements.txt
--index-url https://pypi.python.org/simple/

django~=2.0
git+https://github.com/TransparenCEE/moonsheep#egg=django-moonsheep

# REST
djangorestframework
djangorestframework-jsonapi
django-filter

newrelic
whitenoise==3.3
psycopg2
path.py==2.4.1
ipdb==0.10.2
ipython==5.3.0
gunicorn==19.3.0
dj-database-url==0.4.1
